requirements.txt
garminconnect==0.1.50
flask==2.3.3
pandas==2.1.4
numpy==1.26.2
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-caching==2.1.0
flask-limiter==3.5.0
flask-httpauth==4.8.0
pyjwt==2.8.0
celery==5.3.6
redis==5.0.1
prometheus-client==0.19.0
psutil==5.9.7
gunicorn==21.2.0
python-dotenv==1.0.0
